Checksum calculation and saving implemented
Removed `sudo dnf update` as it is not required and also requires the user to input the `sudo` password, which is inconvenient. Checksum is calculated from the following: the directories of `configs`, `repos`, and `script` in the configuration directory. It is then saved to a checksum file in `~/.local/state/fedorafig` for later comparison. For easy calculation in different subparsers, `Check.calc_checksum` is now static method.