Implemented checksum check and made minor fixes
Checksum is calculated every run and checked with the existing checksum. If they are different or if there is no checksum file, `check` is run. Moved some exceptions into `errors.py`, and fixed `main.py` and `check.py` to be compatible. Currently, `files_do` flag for some reason copies the contents of `cfgpath` to the present working directory.